Partially implement backbone selection
The structure is there but it seems that there are some issues with using
backbones other than ResNet. It seems that they generate too many features.

One change I'm not 100% sure of is in the forward function. It seems that
some models create multi-dimensional feature tensors so I need to use flatten
instead of squeeze.

```
      encodings = [self.fc1(torch.flatten(self.backbone1(imgs[0]), start_dim = 1)),
                   self.fc1(torch.flatten(self.backbone2(imgs[1]), start_dim = 1)),
                   self.fc1(torch.flatten(self.backbone3(imgs[2]), start_dim = 1)),
                   self.fc1(torch.flatten(self.backbone4(imgs[3]), start_dim = 1))]
```