tf Classifier
Logic is similar. The purpose is to rewrite training code and try the tensorflow pertained model just to make sure our models and PyTorch code is not flawed. Compare experiment results to PyTorch results. Some of the functionality is not there yet including freezing the layers. Will add in the next version.